Avoid signed-unsigned comparison
This had the effect of the error-handling path not being used
practically ever, since the signed bytesWritten was coerced into its
large unsigned representation, although that may be compiler-specific
behavior.

+++ pg_logfebe.c
@@ -654,7 +654,7 @@ static void
 sendOrInval(int *fd, char *payload, size_t payloadSz)
 {
 	const int saved_errno = errno;
-	int bytesWritten;
+	ssize_t bytesWritten;
 
 writeAgain:
 	errno = 0;
@@ -669,7 +669,11 @@ writeAgain:
 	 */
 	bytesWritten = send(*fd, payload, payloadSz, MSG_NOSIGNAL);
 
-	if (bytesWritten < payloadSz)
+	/*
+	 * NB: Carefully perform signed-integer conversion to ssize_t;
+	 * otherwise the comparison delivers unintuitive results.
+	 */
+	if (bytesWritten < (ssize_t) payloadSz)
 	{
 		/*
 		 * Something went wrong.
